requirements.txt
brotli
jinja2
psycopg2-binary
requests
pycryptodome